backend/requirements.txt
# Web framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6

# Config
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# LangChain ecosystem (use compatible ranges)
langchain>=0.3.0,<0.4.0
langgraph>=0.2.0,<1.0.0
langchain-community>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0

# Google AI
google-generativeai>=0.8.0

# Vector store & embeddings
chromadb>=0.5.0,<1.0.0
sentence-transformers>=3.0.0,<4.0.0

# Document processing
PyPDF2>=3.0.1,<4.0.0

# Graph database
neo4j>=5.16.0,<6.0.0

# Search
rank-bm25>=0.2.2,<1.0.0

# HTTP client
httpx>=0.26.0,<1.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0


pdfplumber>=0.10.0


